iot_smart-agri/consumer/requirements.txt
kafka-python
paho-mqtt
pymongo